+++ docs/media/AbstractConfigurableClass.ts
@@ -4,18 +4,15 @@
  * @packageDocumentation
  */
 /**
- * @package @maddimathon/utility-typescript@0.4.0-draft
+ * @package @maddimathon/utility-typescript@0.4.0
  */
 /*!
- * @maddimathon/utility-typescript@0.4.0-draft
+ * @maddimathon/utility-typescript@0.4.0
  * @license MIT
  */
 
 import { mergeArgs } from '../../functions/index.js';
-import {
-    // MergeObjects,
-    RecursivePartial,
-} from '../../types/objects/basics.js';
+import { RecursivePartial } from '../../types/objects/basics.js';
 
 
 /**